# Improve FP8 BMM heuristic for large shapes and MoE E2E performance