Summary:
This Diff improves FP8 BMM heuristic large shapes and MoE E2E performance
- Improve FP8 BMM heuristic large shapes with up to 2.5x speedup compared to existing heuristic

- Make bmm torch.compile optional instead of by default, as it recompiles every fp8 rowwise quantization triton call which slows down prefill that lacks cudagraph support. Will revisit it in the future

## Microbenchmark:
 {F1955650829} 

## E2E loadgen
With 17B x 16: input length 2048 (max input length: 8192), output length: 150, batch sizes: 32 (max batch size = 64):

BF16 Trace: https://fburl.com/perfdoctor/uqlwiwre
- TTFT: 289.850 ms
- TTIT: 38.749 ms

FP8 Trace: https://fburl.com/perfdoctor/ruun55sb
- TTFT: 252.600 ms (**1.14x**)
- TTIT: 30.577 ms (**1.26x**)

Differential Revision: D65647034